Animate movement vectors

animate_movement_vectors() produces a gganim object in which each frame is a map of vectors showing the average modeled movement for all birds from each cell in the landscape at a given timestep. It is analogous to a series of images created with plot_movement_vectors().

Usage

animate_movement_vectors(bf, start = "all", ...)

Arguments

bf

a BirdFlow object

start

The starting points in time specified as timesteps, character dates, or date objects; or may be set to "all" or a season name to be interpreted by lookup_season_timesteps().

...

Arguments passed on to lookup_timestep_sequence

end

If start is a timestep or date end should be a timestep or date indicating the ending point in time.

direction

Either "forward" or "backward".

If start and end represent dates and direction is used an error will be thrown if direction isn't consistent with direction indicated by the dates.

If start and end are not dates, direction defaults to "forward" and start and end should either both be timesteps (numeric); or end should be omitted and start should be "all" or a season name.

season_buffer

Only used if start is a season. season_buffer is passed to lookup_season_timesteps() and defaults to 1; it is the number of timesteps to extend the season by at each end.

Value

a gganim object. print() will plot it with default options, or use animate() to set the options. See the example for recommended settings.

Details

Each arrow represents the average of all the transitions from a single cell. The tail of the arrow is the center of that cell, the head is the average location at the following timestep for birds that start at that cell.

The timestep and/or date label is the starting timestep for the transition that is displayed and the format depends on birdflow_options("time_format")

Thicker lines and less transparency (darker shading) indicate higher density in the eBird S&T distribution for the beginning timestep of the displayed transition.

I'm undecided as to whether the gganimate() call from the example should be part of the function, making it return a "gif_image" object instead. It would make it easier to use but hard to adjust the rendering parameters. The output object might change in a future revision.

Examples

bf <- BirdFlowModels::amewoo
a <- animate_movement_vectors(bf)
#> Creating vector fields
#> 	...................................................

if (FALSE) {

# Animate, display, and save
#   Note: "ragg_png" is considerably faster and produces cleaner output than
#         the default device.
gif <- gganimate::animate(a, fps = 1, device = "ragg_png",
                          width = 6, height = 5,
                          res = 150, units = "in")
print(gif)
path <- tempfile(fileext = ".gif") # set to desired output location
gganimate::anim_save(gif)
}