Define the coordinate reference system

This defines the coordinate reference system (CRS, AKA projection) used by default for BirdFlow models. It is a customized Mollweide projection, with the longitude of origin set to -90 deg. centering the western hemisphere. St Louis, Missouri is very close to a longitude of -90 deg. The Mollweide projection preserves area, shape distortion increases with distance from the longitude of origin.

Usage

birdflow_crs

Format

This is a string defining a custom Mollweide projection centered on the western hemisphere with well known text.

Details

The projection is similar to: SR-ORG:7399 and ESRI:54009 which have a longitude of origin of 0, centered on Greenwich, England.

References

https://epsg.io/54009

Examples

cat(birdflow_crs)
#> PROJCRS["Western Mollweide",
#>     BASEGEOGCRS["WGS 84",
#>         DATUM["World Geodetic System 1984",
#>             ELLIPSOID["WGS 84",6378137,298.257223563,
#>                 LENGTHUNIT["metre",1]]],
#>         PRIMEM["Greenwich",0,
#>             ANGLEUNIT["Degree",0.0174532925199433]]],
#>     CONVERSION["Western Mollweide",
#>         METHOD["Mollweide"],
#>         PARAMETER["Longitude of natural origin",-90,
#>             ANGLEUNIT["Degree",0.0174532925199433],
#>             ID["EPSG",8802]],
#>         PARAMETER["False easting",0,
#>             LENGTHUNIT["metre",1],
#>             ID["EPSG",8806]],
#>         PARAMETER["False northing",0,
#>             LENGTHUNIT["metre",1],
#>             ID["EPSG",8807]]],
#>     CS[Cartesian,2],
#>         AXIS["(E)",east,
#>             ORDER[1],
#>             LENGTHUNIT["metre",1]],
#>         AXIS["(N)",north,
#>             ORDER[2],
#>             LENGTHUNIT["metre",1]],
#>     USAGE[
#>         SCOPE["Not known."],
#>         AREA["World."],
#>         BBOX[-90,-180,90,180]]]
crs(birdflow_crs, proj = TRUE)
#> [1] "+proj=moll +lon_0=-90 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs"