Abundance Estimates from Callaghan et. al. 2021

The Callaghan et. al. (2021) population estimates and 95% confidence interval for 9700 species.

Usage

callaghan_abundance

Format

A data frame with 9,700 rows and 11 columns:

common_name

Common name

scientific_name

Scientific name

order, family

Taxonomic order and family

lower_ci

The lower 95% confidence interval

abundance_estimate

Estimated population size

upper_ci

The upper 95% confidence interval

range_adjusted

TRUE indicates that the area of the species range from a source other than the eBird pixels was used to adjust the abundance estimate. If an independent range area was not available than the area of the eBird cells was used. See Callaghan et. al. 2021"

training_species

TRUE if the species was used to train the abundance model (Callaghan et. al. 2021)

species_code

The eBird species code derived by joining scientific_name to the ebird_taxonomy in the auk package. NA indicates that the scientific name from Callaghan et. al. did not match any scientific names in the eBird taxonomy table.

in_ebird_2022

Is the species_code in ebirds::ebirdst_runs from the 2022 version of the eBird package. See: ebirdst::ebirdst_version() and ebirdst::ebirdst_runs. Starting with the 2022 version year only a subset of the species were fit by eBird in each year.

Source

Callaghan, Corey T., Shinichi Nakagawa, and William K. Cornwell. "Global abundance estimates for 9,700 bird species." Proceedings of the National Academy of Sciences 118.21 (2021): e2023170118. https://doi.org/10.1073/pnas.2023170118