Function to convert a raster bird distribution into its collapsed, vector equivalent

This function is primarily for internal use. It converts between the raster representation of data in which rows and columns indicate position in space and a vector representation that contains all the non-masked cells in row major order - starting at the top left and proceeding left to right along each row. The collapsed form is used for projecting the flow model while the expanded form is used to export, import, and visualize the data.

Usage

collapse_distr(x, bf)

Arguments

x

Either a matrix representing a single bird distribution or an array representing multiple distributions.

bf

A BirdFlow model or the geom component of one.

Value

Either a vector representing a single distribution in its collapsed form or, if x represents multiple distributions, a matrix with one distribution per column.

See also

Examples

if (FALSE) {
# Requires BirdFlow object 'bf'
d <- get_distr(1, bf) # compacted form
e <- expand_distr(d, bf) # expand
f <- collapse_distr(e, bf)
stopifnot(all(d == f))
}