retrieve BirdFlow metadata

get_metadata retreives information about the BirdFlow model: when it was created, what version of eBird Status & Trends data was used, whether it has been sparsified, etc.

Usage

get_metadata(x, what)

Arguments

x

BirdFlow model

what

Indicates what metadata to return. The defualt "all" returns the complete list. Options are described in details.

Value

Metadata from the BirdFlow model. Potentially a list.

Details

what may be used with these options:

all	Return the complete list, this is the default

has_marginals	Logical, indicates whether the model has marginals. Marginals encode forward and backward transitions, and distributions in an n by n matrix for each pair of adjacent timesteps.

has_transitions	Logical, indicates whether the model has stored transitions (separate from marginal)

has_distr	Logical, indicates whether the model has stored distributions (separate from marginals); if it does they are from ebirdst and are the distributions the model was trained on

n_transitions	Number of transitions in the model

n_active	The number of active states in the model. This defines the dimensions of the marginals and transitions matrices and the number of columns in flattened distributions. In a full (not sparse) model all active states could be visited.

n_timesteps	The number of timesteps in the model.

ebird_version_year	The version year for the eBird Status and Trends Data the model was trained on see ebirdst_version()

ebird_release_year	The year the eBird S & T data was released ebirdst_version()

ebird_access_end_date	The date at which the training eBirds S & T data will no longer be available. This does not prevent continued use of the BirdFlow model

birdflow_preprocess_date	The date the data was downloaded with ebirdst and formated for BirdFlow model fitting (see preprocess_species()

birdflow_model_date	The date the model was fitted and/or exported to hdf5 from python"

is_sparse	Logical, if TRUE the model is sparse. See sparsify()

sparse_stats	If the model is sparse this contains sparsification statistics (a list)

See also

Examples

library(BirdFlowModels)
get_metadata(amewoo)
#> $has_marginals
#> [1] TRUE
#> 
#> $has_transitions
#> [1] FALSE
#> 
#> $has_distr
#> [1] TRUE
#> 
#> $n_transitions
#> [1] 51
#> 
#> $n_active
#> [1] 1706
#> 
#> $n_timesteps
#> [1] 52
#> 
#> $ebird_version_year
#> [1] NA
#> 
#> $ebird_release_year
#> [1] NA
#> 
#> $ebird_access_end_date
#> [1] NA
#> 
#> $birdflow_preprocess_date
#> [1] NA
#> 
#> $birdflow_model_date
#> [1] "2023-01-02 10:03:05.256454"
#> 
#> $birdflow_version
#> [1] 2
#> 
#> $is_sparse
#> [1] TRUE
#> 
#> $sparse
#> $sparse$fix_stats
#>   step pct_lost n_dead_ends
#> 1    0 6.354389         172
#> 2    1 6.376854          83
#> 3    2 6.383977          37
#> 4    3 6.385986           9
#> 5    4 6.386376           1
#> 6    5 6.386384           0
#> 
#> $sparse$method
#> [1] "conditional" "state"      
#> 
#> $sparse$arguments
#> $sparse$arguments$p
#> [1] 0.99
#> 
#> 
#> $sparse$stats
#>    model pct_zero pct_density_lost mean_step_cor min_step_cor traverse_cor
#> 1   full  0.00000         0.000000     0.9947801    0.9831586    0.9484126
#> 2 sparse 98.94806         6.354389     0.9945607    0.9823753    0.9465323
#> 3  fixed 98.95086         6.386384     0.9945142    0.9823376    0.9464339
#>   mean_distr_cor min_distr_cor
#> 1      0.9919662     0.9689286
#> 2      0.9927087     0.9685167
#> 3      0.9927038     0.9684455
#> 
#> $sparse$pct_zero
#> [1] 98.95086
#> 
#> $sparse$pct_density_lost
#> [1] 6.386384
#> 
#> 
get_metadata(amewoo, "is_sparse")
#> [1] TRUE