Return the population estimate for a Bird Flow model or species

Return the population estimate for a Bird Flow model or species

Usage

get_population(x, what = "abundance")

Arguments

x

Either a single BirdFlow model, or one or more species eBird codes, scientific names, or common names.

what

Indicates what information will be returned. Options are:

"df"

A data frame with relevant rows from callaghan_abundance

"abundance"

The abundance estimate.

"lower_ci"

The lower 95 percent confidence interval

"upper_ci"

The upper 95 percent confidence interval

Value

See what. In all cases the values correspond to x and will be NA when x cannot be cross walked to callaghan_abundance.

References

Callaghan, Corey T., Shinichi Nakagawa, and William K. Cornwell. "Global abundance estimates for 9,700 bird species." Proceedings of the National Academy of Sciences 118.21 (2021): e2023170118. https://doi.org/10.1073/pnas.2023170118

Examples

bf <- BirdFlowModels::amewoo
get_population(bf)
#> [1] 698716
get_population("American Black Duck")
#> [1] 18953154
get_population("amewoo")
#> [1] 698716