Lookup eBird species metadata

lookup_species_metadata() uses ebirdst to generate a list identical to the species component of a BirdFlow model. It is an internal function used by preprocess_species() and Routes(). See species_info() for a description of the list items.

Usage

lookup_species_metadata(
  species,
  skip_checks = FALSE,
  min_season_quality = 3,
  quiet = FALSE
)

Arguments

species

An eBird species code, common name, or scientific name. It will be processed by ebirdst::get_species().

skip_checks

Set to TRUE to skip the checks for eBird model quality and for migrant status. Defaults to TRUE - do the tests.

min_season_quality

If skip_checks = FALSE and the model quality for any of the four seasons drops below this threshold an error is thrown.

quiet

Set to TRUE to suppress messages.

Value

A list of species information derived from ebirdst. See species_info() for a description of the items.

See also