Retrieve BirdFlow metadata

get_metadata retrieves information about the BirdFlow model: when it was created, what version of eBird Status & Trends data was used, whether it has been sparsified, etc.

Usage

get_metadata(x, what)

Arguments

x

BirdFlow model

what

Indicates what metadata to return. The default "all" returns the complete list. Options are described in details.

Value

Metadata from the BirdFlow model. Potentially a list.

Details

what may be used with these options:

all	Return the complete list, this is the default

has_marginals	Logical, indicates whether the model has marginals. Marginals encode forward and backward transitions, and distributions in an n by n matrix for each pair of adjacent timesteps.

has_transitions	Logical, indicates whether the model has stored transitions (separate from marginal)

has_distr	Logical, indicates whether the model has stored distributions (separate from marginals); if it does they are from ebirdst and are the distributions the model was trained on

n_transitions	Number of transitions in the model.

n_active	The number of active states in the model. This defines the dimensions of the marginals and transitions matrices and the number of columns in flattened distributions. In a full (not sparse) model all active states could be visited.

n_timesteps	The number of timesteps in the model.

ebird_version_year	The version year for the eBird Status and Trends Data the model was trained on see ebirdst_version()

ebird_release_year	The year the eBird S & T data was released ebirdst_version()

ebird_access_end_date	The date at which the training eBirds S & T data will no longer be available. This does not prevent continued use of the BirdFlow model.

birdflow_preprocess_date	The date the data was downloaded with ebirdst and formatted for BirdFlow model fitting (see preprocess_species()

birdflow_model_date	The date the model was fitted and/or exported to hdf5 from Python.

is_sparse	Logical, if TRUE the model is sparse. See sparsify()

sparse_stats	If the model is sparse this contains sparsification statistics (a list).

See also

Examples

library(BirdFlowModels)
get_metadata(amewoo)
#> $has_marginals
#> [1] TRUE
#> 
#> $has_transitions
#> [1] FALSE
#> 
#> $has_distr
#> [1] TRUE
#> 
#> $n_transitions
#> [1] 52
#> 
#> $n_active
#> [1] 342
#> 
#> $n_timesteps
#> [1] 52
#> 
#> $timestep_padding
#> [1] 2
#> 
#> $ebird_version_year
#> [1] 2021
#> 
#> $ebird_release_year
#> [1] 2022
#> 
#> $ebird_access_end_date
#> [1] "2023-11-30"
#> 
#> $birdflow_preprocess_date
#> [1] "2023-11-21"
#> 
#> $birdflow_model_date
#> [1] "2023-11-21 17:19:27.009766"
#> 
#> $birdflow_version
#> [1] 3
#> 
#> $birdflowr_version
#> [1] "0.1.0.9038"
#> 
#> $is_sparse
#> [1] FALSE
#> 
#> $sparse
#> [1] NA
#> 
#> $hyperparameters
#> $hyperparameters$dist_pow
#> [1] 0.6
#> 
#> $hyperparameters$dist_weight
#> [1] 0.008658
#> 
#> $hyperparameters$ent_weight
#> [1] 0.001443
#> 
#> $hyperparameters$learning_rate
#> [1] 0.1
#> 
#> $hyperparameters$normalized
#> [1] TRUE
#> 
#> $hyperparameters$obs_weight
#> [1] 1
#> 
#> $hyperparameters$rng_seed
#> [1] 17
#> 
#> $hyperparameters$training_steps
#> [1] 600
#> 
#> 
#> $loss_values
#>          dist       ent        obs     total
#> 1   215.15869 236.17322 0.57939780 2.1014438
#> 2   210.01872 236.22165 0.52866054 2.0061350
#> 3   204.95619 235.81674 0.48049927 1.9147265
#> 4   200.01207 234.99609 0.43503928 1.8276443
#> 5   195.21609 233.80389 0.39228603 1.7450881
#> 6   190.58633 232.28596 0.35218555 1.6670935
#> 7   186.13089 230.49072 0.31466532 1.5935885
#> 8   181.84903 228.46327 0.27965516 1.5244315
#> 9   177.73326 226.24446 0.24709456 1.4594384
#> 10  173.77129 223.86955 0.21693729 1.3984053
#> 11  169.94862 221.36865 0.18916348 1.3411437
#> 12  166.25012 218.76643 0.16378160 1.2874954
#> 13  162.66167 216.08530 0.14081964 1.2373333
#> 14  159.17067 213.34410 0.12031531 1.1905594
#> 15  155.76617 210.56020 0.10230915 1.1470944
#> 16  152.43875 207.75093 0.08682334 1.1068535
#> 17  149.17995 204.93588 0.07382077 1.0696983
#> 18  145.98276 202.13470 0.06316140 1.0353997
#> 19  142.84262 199.36826 0.05458094 1.0036240
#> 20  139.75818 196.65489 0.04771935 0.9739727
#> 21  136.73073 194.00783 0.04219241 0.9460538
#> 22  133.76320 191.43358 0.03766023 0.9195434
#> 23  130.85825 188.93268 0.03386035 0.8942012
#> 24  128.01682 186.50026 0.03061645 0.8698661
#> 25  125.23743 184.12532 0.02783360 0.8464464
#> 26  122.51659 181.79424 0.02547722 0.8238968
#> 27  119.84933 179.49156 0.02354145 0.8021907
#> 28  117.23013 177.20245 0.02201207 0.7812875
#> 29  114.65356 174.91286 0.02084434 0.7611157
#> 30  112.11458 172.61061 0.01996790 0.7415788
#> 31  109.60916 170.28621 0.01930417 0.7225773
#> 32  107.13448 167.93297 0.01878082 0.7040239
#> 33  104.68911 165.54652 0.01833465 0.6858493
#> 34  102.27284 163.12595 0.01790984 0.6679974
#> 35   99.88613 160.67148 0.01746781 0.6504330
#> 36   97.52901 158.18375 0.01700313 0.6331502
#> 37   95.20029 155.66302 0.01654312 0.6161655
#> 38   92.89687 153.10956 0.01612644 0.5994905
#> 39   90.61395 150.52261 0.01578162 0.5831131
#> 40   88.34538 147.89935 0.01552618 0.5670018
#> 41   86.08443 145.23489 0.01537028 0.5511154
#> 42   83.82442 142.52255 0.01530915 0.5354009
#> 43   81.55951 139.75478 0.01533122 0.5198073
#> 44   79.28568 136.92514 0.01544342 0.5043159
#> 45   77.00089 134.02835 0.01565972 0.4889304
#> 46   74.70551 131.06082 0.01595673 0.4736363
#> 47   72.40195 128.02028 0.01627936 0.4584022
#> 48   70.09305 124.90343 0.01659533 0.4432253
#> 49   67.78003 121.70702 0.01690826 0.4281245
#> 50   65.46175 118.42944 0.01723262 0.4131068
#> 51   63.13724 115.07220 0.01758363 0.3981767
#> 52   60.80972 111.64002 0.01797476 0.3833687
#> 53   58.48853 108.14057 0.01839230 0.3687391
#> 54   56.18575 104.58351 0.01877982 0.3543221
#> 55   53.90903 100.97726 0.01912162 0.3401558
#> 56   51.65814  97.32857 0.01948931 0.3263004
#> 57   49.43314  93.64602 0.01993422 0.3127952
#> 58   47.24622  89.94505 0.02043025 0.2996973
#> 59   45.11935  86.25188 0.02088596 0.2870678
#> 60   43.06939  82.59638 0.02124022 0.2749484
#> 61   41.09838  78.99474 0.02156441 0.2634048
#> 62   39.20367  75.45681 0.02193111 0.2524723
#> 63   37.39429  71.99626 0.02231174 0.2421809
#> 64   35.68283  68.63718 0.02265259 0.2325511
#> 65   34.07260  65.39851 0.02296549 0.2235960
#> 66   32.56284  62.28904 0.02327248 0.2153185
#> 67   31.15820  59.31972 0.02353251 0.2077019
#> 68   29.86086  56.50695 0.02373124 0.2007270
#> 69   28.66398  53.85399 0.02391851 0.1943800
#> 70   27.56353  51.35767 0.02409757 0.1886335
#> 71   26.55832  49.02469 0.02425246 0.1834518
#> 72   25.64515  46.85950 0.02438642 0.1788039
#> 73   24.82282  44.85614 0.02447482 0.1746634
#> 74   24.08561  43.01480 0.02451892 0.1709818
#> 75   23.42373  41.33087 0.02453936 0.1677016
#> 76   22.83664  39.79052 0.02448357 0.1647855
#> 77   22.31847  38.39343 0.02437053 0.1622021
#> 78   21.85636  37.12669 0.02425627 0.1599148
#> 79   21.45346  35.97917 0.02405668 0.1578828
#> 80   21.08941  34.93098 0.02389091 0.1560776
#> 81   20.77568  33.99057 0.02364965 0.1544771
#> 82   20.48409  33.12300 0.02350583 0.1530606
#> 83   20.24711  32.37160 0.02325953 0.1518468
#> 84   19.99974  31.65578 0.02329766 0.1507761
#> 85   19.82391  31.06334 0.02288498 0.1496960
#> 86   19.63743  30.50491 0.02269227 0.1486946
#> 87   19.47584  29.99916 0.02260628 0.1479394
#> 88   19.35129  29.56940 0.02230296 0.1471778
#> 89   19.21871  29.15513 0.02208602 0.1464108
#> 90   19.10439  28.77748 0.02194645 0.1458264
#> 91   19.01208  28.44985 0.02167320 0.1452267
#> 92   18.91162  28.12768 0.02146711 0.1446157
#> 93   18.81471  27.82571 0.02140362 0.1441489
#> 94   18.73859  27.56047 0.02117302 0.1436420
#> 95   18.66191  27.30120 0.02096670 0.1431458
#> 96   18.57918  27.05175 0.02091661 0.1427395
#> 97   18.52125  26.83230 0.02067105 0.1423090
#> 98   18.45988  26.62218 0.02049173 0.1419016
#> 99   18.38980  26.41711 0.02044983 0.1415489
#> 100  18.34637  26.23856 0.02020362 0.1411843
#> 101  18.29744  26.06705 0.02004214 0.1408466
#> 102  18.23352  25.89383 0.02003081 0.1405318
#> 103  18.19164  25.74530 0.01987303 0.1402258
#> 104  18.16162  25.60646 0.01965370 0.1399469
#> 105  18.09734  25.45772 0.01971973 0.1396710
#> 106  18.07183  25.33311 0.01949974 0.1394099
#> 107  18.03215  25.21040 0.01943330 0.1391770
#> 108  17.98865  25.08531 0.01940238 0.1389500
#> 109  17.95098  24.98006 0.01934366 0.1387170
#> 110  17.93885  24.88687 0.01910933 0.1385121
#> 111  17.88954  24.78293 0.01917116 0.1382970
#> 112  17.86962  24.70027 0.01902344 0.1380961
#> 113  17.84707  24.62154 0.01892032 0.1379114
#> 114  17.81101  24.53426 0.01891677 0.1377216
#> 115  17.79047  24.46457 0.01881874 0.1375462
#> 116  17.76850  24.39696 0.01874305 0.1373779
#> 117  17.74545  24.32319 0.01867010 0.1372119
#> 118  17.71063  24.25842 0.01871888 0.1370527
#> 119  17.71116  24.20341 0.01848524 0.1369029
#> 120  17.66461  24.12589 0.01862474 0.1367512
#> 121  17.66088  24.07239 0.01842345 0.1365949
#> 122  17.63447  24.01230 0.01841966 0.1364492
#> 123  17.61185  23.94631 0.01838547 0.1363143
#> 124  17.59656  23.89375 0.01831144 0.1361838
#> 125  17.57432  23.84183 0.01830724 0.1360619
#> 126  17.55869  23.78236 0.01822589 0.1359311
#> 127  17.54073  23.73050 0.01817656 0.1358011
#> 128  17.52838  23.68481 0.01810650 0.1356900
#> 129  17.50441  23.62617 0.01812241 0.1355830
#> 130  17.49388  23.57967 0.01803668 0.1354733
#> 131  17.47589  23.53364 0.01800993 0.1353572
#> 132  17.45754  23.48322 0.01797747 0.1352385
#> 133  17.44602  23.43928 0.01790741 0.1351322
#> 134  17.43336  23.39745 0.01786024 0.1350357
#> 135  17.41234  23.34965 0.01786687 0.1349294
#> 136  17.40497  23.30858 0.01776393 0.1348219
#> 137  17.38550  23.26701 0.01777706 0.1347264
#> 138  17.37332  23.22565 0.01772797 0.1346315
#> 139  17.36081  23.18967 0.01769353 0.1345407
#> 140  17.34868  23.15325 0.01766283 0.1344576
#> 141  17.33361  23.11499 0.01766050 0.1343800
#> 142  17.33155  23.09148 0.01757074 0.1343063
#> 143  17.31046  23.05419 0.01763810 0.1342448
#> 144  17.31555  23.03251 0.01751948 0.1342016
#> 145  17.28115  22.99633 0.01772503 0.1341615
#> 146  17.30247  22.98747 0.01745017 0.1340841
#> 147  17.26515  22.94857 0.01759075 0.1339576
#> 148  17.26840  22.93759 0.01744519 0.1338560
#> 149  17.26815  22.92123 0.01737379 0.1338061
#> 150  17.23917  22.88730 0.01753343 0.1337638
#> 151  17.25380  22.88193 0.01731527 0.1336800
#> 152  17.23484  22.85383 0.01733632 0.1335775
#> 153  17.22331  22.82814 0.01733726 0.1335157
#> 154  17.23231  22.81803 0.01720790 0.1334788
#> 155  17.20709  22.78728 0.01731037 0.1334073
#> 156  17.20583  22.76919 0.01720963 0.1333218
#> 157  17.20079  22.75128 0.01716987 0.1332642
#> 158  17.18292  22.72430 0.01724168 0.1332203
#> 159  17.18347  22.70880 0.01715875 0.1331644
#> 160  17.16820  22.68656 0.01718856 0.1330941
#> 161  17.16698  22.67139 0.01711558 0.1330325
#> 162  17.16062  22.65629 0.01709819 0.1329818
#> 163  17.14534  22.63556 0.01714992 0.1329312
#> 164  17.15091  22.62564 0.01702911 0.1328728
#> 165  17.13435  22.60651 0.01708022 0.1328082
#> 166  17.12914  22.58965 0.01704250 0.1327498
#> 167  17.12807  22.57579 0.01698223 0.1327002
#> 168  17.11175  22.55549 0.01704234 0.1326483
#> 169  17.11417  22.54078 0.01694166 0.1325898
#> 170  17.09871  22.52135 0.01698649 0.1325288
#> 171  17.09552  22.50710 0.01693641 0.1324717
#> 172  17.09214  22.49036 0.01688941 0.1324196
#> 173  17.07745  22.47151 0.01693996 0.1323702
#> 174  17.07973  22.46056 0.01686089 0.1323266
#> 175  17.06517  22.43611 0.01691041 0.1322854
#> 176  17.06181  22.42456 0.01687788 0.1322404
#> 177  17.05326  22.40639 0.01687718 0.1321918
#> 178  17.04973  22.39306 0.01684198 0.1321454
#> 179  17.04434  22.37966 0.01682299 0.1320990
#> 180  17.03833  22.36742 0.01681275 0.1320544
#> 181  17.03879  22.35432 0.01674982 0.1320143
#> 182  17.02162  22.33428 0.01683128 0.1319761
#> 183  17.02952  22.32810 0.01671239 0.1319346
#> 184  17.01516  22.30757 0.01675859 0.1318860
#> 185  17.01153  22.29440 0.01672478 0.1318398
#> 186  17.00545  22.27884 0.01670907 0.1317939
#> 187  16.99896  22.26471 0.01670023 0.1317493
#> 188  16.99821  22.25034 0.01664553 0.1317088
#> 189  16.98424  22.23098 0.01670326 0.1316735
#> 190  16.98877  22.22224 0.01661739 0.1316395
#> 191  16.97197  22.19886 0.01669724 0.1316076
#> 192  16.98023  22.19510 0.01658640 0.1315737
#> 193  16.96098  22.17628 0.01668637 0.1315342
#> 194  16.97062  22.17128 0.01654852 0.1314870
#> 195  16.95835  22.15936 0.01659032 0.1314397
#> 196  16.95742  22.15170 0.01654929 0.1314017
#> 197  16.95791  22.14238 0.01649891 0.1313690
#> 198  16.94660  22.12874 0.01654701 0.1313389
#> 199  16.95179  22.12420 0.01646564 0.1313090
#> 200  16.93775  22.10438 0.01652597 0.1312764
#> 201  16.94481  22.10275 0.01642388 0.1312378
#> 202  16.93049  22.08532 0.01648117 0.1311963
#> 203  16.93337  22.07756 0.01640686 0.1311581
#> 204  16.92687  22.06677 0.01641270 0.1311232
#> 205  16.92020  22.05248 0.01641692 0.1310903
#> 206  16.92011  22.04379 0.01637454 0.1310597
#> 207  16.91204  22.03129 0.01639866 0.1310320
#> 208  16.91214  22.02149 0.01635873 0.1310070
#> 209  16.89930  22.01009 0.01642817 0.1309817
#> 210  16.90663  22.00468 0.01632975 0.1309546
#> 211  16.89034  21.99017 0.01642132 0.1309261
#> 212  16.89895  21.98959 0.01631496 0.1308951
#> 213  16.88288  21.97270 0.01639443 0.1308598
#> 214  16.88880  21.97260 0.01630564 0.1308224
#> 215  16.88251  21.96353 0.01631340 0.1307888
#> 216  16.87732  21.95284 0.01631655 0.1307625
#> 217  16.88304  21.95287 0.01624608 0.1307415
#> 218  16.86899  21.93585 0.01632453 0.1307228
#> 219  16.87714  21.93240 0.01623304 0.1307068
#> 220  16.85947  21.91680 0.01634695 0.1306903
#> 221  16.86988  21.91494 0.01622834 0.1306645
#> 222  16.85229  21.89596 0.01631272 0.1306240
#> 223  16.85730  21.89618 0.01622939 0.1305837
#> 224  16.85430  21.88583 0.01621110 0.1305543
#> 225  16.84208  21.87434 0.01628146 0.1305355
#> 226  16.85350  21.87514 0.01616985 0.1305216
#> 227  16.83553  21.85907 0.01628710 0.1305065
#> 228  16.84516  21.86051 0.01617789 0.1304785
#> 229  16.83455  21.84846 0.01621572 0.1304419
#> 230  16.83792  21.84616 0.01615122 0.1304099
#> 231  16.83299  21.83891 0.01616018 0.1303867
#> 232  16.82916  21.82931 0.01616184 0.1303690
#> 233  16.83392  21.82792 0.01610274 0.1303531
#> 234  16.82036  21.81409 0.01618327 0.1303362
#> 235  16.82716  21.80908 0.01609654 0.1303156
#> 236  16.81250  21.79788 0.01618094 0.1302892
#> 237  16.81925  21.79479 0.01609167 0.1302629
#> 238  16.80925  21.78427 0.01614055 0.1302403
#> 239  16.80967  21.78138 0.01611381 0.1302214
#> 240  16.80852  21.77526 0.01609932 0.1302058
#> 241  16.80143  21.76492 0.01613048 0.1301904
#> 242  16.80258  21.76468 0.01610328 0.1301736
#> 243  16.79903  21.75423 0.01609475 0.1301494
#> 244  16.79503  21.75133 0.01609648 0.1301206
#> 245  16.79674  21.74868 0.01604964 0.1300925
#> 246  16.79145  21.74013 0.01606167 0.1300711
#> 247  16.79146  21.73843 0.01604202 0.1300539
#> 248  16.78631  21.73178 0.01605991 0.1300369
#> 249  16.78663  21.72808 0.01603318 0.1300182
#> 250  16.77645  21.71863 0.01608490 0.1299954
#> 251  16.78213  21.71862 0.01601093 0.1299706
#> 252  16.77182  21.70655 0.01605967 0.1299476
#> 253  16.77413  21.70441 0.01601765 0.1299286
#> 254  16.77252  21.70133 0.01601220 0.1299137
#> 255  16.77052  21.69327 0.01600377 0.1298995
#> 256  16.76756  21.69035 0.01600979 0.1298842
#> 257  16.77250  21.68781 0.01594688 0.1298677
#> 258  16.75853  21.67697 0.01603410 0.1298496
#> 259  16.76795  21.67728 0.01593725 0.1298339
#> 260  16.75224  21.66353 0.01604327 0.1298237
#> 261  16.76273  21.66734 0.01595336 0.1298191
#> 262  16.74409  21.65052 0.01608374 0.1298124
#> 263  16.76005  21.65949 0.01594011 0.1297940
#> 264  16.74089  21.64371 0.01604921 0.1297600
#> 265  16.75217  21.64893 0.01592282 0.1297237
#> 266  16.74335  21.64136 0.01596323 0.1296987
#> 267  16.74063  21.63709 0.01596956 0.1296876
#> 268  16.74634  21.64055 0.01591945 0.1296820
#> 269  16.73415  21.63003 0.01599761 0.1296697
#> 270  16.74417  21.63561 0.01589780 0.1296486
#> 271  16.73157  21.62630 0.01597012 0.1296253
#> 272  16.73825  21.62832 0.01589629 0.1296064
#> 273  16.73118  21.62293 0.01593526 0.1295919
#> 274  16.73277  21.62078 0.01590365 0.1295772
#> 275  16.72900  21.61789 0.01591292 0.1295580
#> 276  16.72684  21.61219 0.01590157 0.1295362
#> 277  16.72807  21.61094 0.01586993 0.1295170
#> 278  16.72005  21.60370 0.01591609 0.1295041
#> 279  16.72536  21.60117 0.01585618 0.1294939
#> 280  16.71494  21.59375 0.01592314 0.1294814
#> 281  16.72028  21.59209 0.01585807 0.1294649
#> 282  16.71193  21.58495 0.01590064 0.1294454
#> 283  16.71602  21.58465 0.01584640 0.1294271
#> 284  16.70961  21.57806 0.01587757 0.1294122
#> 285  16.71255  21.57806 0.01584084 0.1294009
#> 286  16.70636  21.57059 0.01587228 0.1293895
#> 287  16.70914  21.57027 0.01583366 0.1293755
#> 288  16.70313  21.56313 0.01585849 0.1293586
#> 289  16.70409  21.56218 0.01583145 0.1293413
#> 290  16.70119  21.55728 0.01583365 0.1293254
#> 291  16.70082  21.55481 0.01581962 0.1293118
#> 292  16.69995  21.55188 0.01581125 0.1293001
#> 293  16.69917  21.54791 0.01580149 0.1292892
#> 294  16.69536  21.54483 0.01581893 0.1292782
#> 295  16.69801  21.54269 0.01578064 0.1292659
#> 296  16.69113  21.53888 0.01582064 0.1292518
#> 297  16.69634  21.53870 0.01576094 0.1292375
#> 298  16.68667  21.53197 0.01582325 0.1292258
#> 299  16.69482  21.53525 0.01575188 0.1292203
#> 300  16.67891  21.52321 0.01587478 0.1292228
#> 301  16.69631  21.53326 0.01574766 0.1292318
#> 302  16.66982  21.51483 0.01595849 0.1292399
#> 303  16.69653  21.53139 0.01574208 0.1292309
#> 304  16.66680  21.51040 0.01593725 0.1291989
#> 305  16.68634  21.52216 0.01574137 0.1291553
#> 306  16.67380  21.51309 0.01580442 0.1291228
#> 307  16.67306  21.51070 0.01579618 0.1291116
#> 308  16.67952  21.51462 0.01574691 0.1291126
#> 309  16.66536  21.50353 0.01584991 0.1291090
#> 310  16.67986  21.51245 0.01572169 0.1290935
#> 311  16.66367  21.50007 0.01582176 0.1290712
#> 312  16.67492  21.50555 0.01571314 0.1290521
#> 313  16.66874  21.50148 0.01574676 0.1290381
#> 314  16.66508  21.49672 0.01576149 0.1290280
#> 315  16.67348  21.50196 0.01568471 0.1290164
#> 316  16.65822  21.49040 0.01578822 0.1290045
#> 317  16.67017  21.49629 0.01568068 0.1289919
#> 318  16.65798  21.48734 0.01576104 0.1289796
#> 319  16.66170  21.48703 0.01571462 0.1289658
#> 320  16.65843  21.48438 0.01572138 0.1289481
#> 321  16.65466  21.47960 0.01572897 0.1289299
#> 322  16.65840  21.48081 0.01568448 0.1289161
#> 323  16.65056  21.47425 0.01573493 0.1289081
#> 324  16.65534  21.47464 0.01568674 0.1289008
#> 325  16.64701  21.46868 0.01573935 0.1288899
#> 326  16.65335  21.47101 0.01567262 0.1288747
#> 327  16.64206  21.46199 0.01574002 0.1288573
#> 328  16.64783  21.46452 0.01567954 0.1288431
#> 329  16.64360  21.46009 0.01569919 0.1288326
#> 330  16.64211  21.45774 0.01570014 0.1288240
#> 331  16.64355  21.45835 0.01567904 0.1288145
#> 332  16.63982  21.45521 0.01569495 0.1288027
#> 333  16.64235  21.45503 0.01565941 0.1287892
#> 334  16.63905  21.45423 0.01567458 0.1287770
#> 335  16.64141  21.45562 0.01564587 0.1287667
#> 336  16.63815  21.45375 0.01566310 0.1287584
#> 337  16.64194  21.45838 0.01563008 0.1287515
#> 338  16.63461  21.45243 0.01567840 0.1287450
#> 339  16.64147  21.45868 0.01562083 0.1287378
#> 340  16.63105  21.45050 0.01569114 0.1287297
#> 341  16.64024  21.45712 0.01561434 0.1287229
#> 342  16.62627  21.44767 0.01571670 0.1287180
#> 343  16.63982  21.45624 0.01561014 0.1287163
#> 344  16.62113  21.44417 0.01575197 0.1287138
#> 345  16.63942  21.45569 0.01560269 0.1287062
#> 346  16.61917  21.44172 0.01574150 0.1286899
#> 347  16.63294  21.45110 0.01561236 0.1286664
#> 348  16.62129  21.44219 0.01567744 0.1286435
#> 349  16.62366  21.44279 0.01564188 0.1286276
#> 350  16.62465  21.44316 0.01562645 0.1286202
#> 351  16.61837  21.43775 0.01567041 0.1286176
#> 352  16.62505  21.44022 0.01561338 0.1286148
#> 353  16.61437  21.43342 0.01568954 0.1286083
#> 354  16.62521  21.43895 0.01559269 0.1285974
#> 355  16.61172  21.42999 0.01568425 0.1285851
#> 356  16.62317  21.43710 0.01558466 0.1285743
#> 357  16.61234  21.42907 0.01565791 0.1285654
#> 358  16.61884  21.43491 0.01560176 0.1285571
#> 359  16.61424  21.42974 0.01562317 0.1285462
#> 360  16.61404  21.42969 0.01561140 0.1285327
#> 361  16.61351  21.42776 0.01559961 0.1285191
#> 362  16.61222  21.42596 0.01559740 0.1285083
#> 363  16.61158  21.42476 0.01559377 0.1285009
#> 364  16.60978  21.42142 0.01559797 0.1284943
#> 365  16.61171  21.42253 0.01557653 0.1284880
#> 366  16.60471  21.41504 0.01561822 0.1284799
#> 367  16.61180  21.41859 0.01555296 0.1284709
#> 368  16.60028  21.40889 0.01562910 0.1284613
#> 369  16.60995  21.41364 0.01554420 0.1284533
#> 370  16.59702  21.40318 0.01563661 0.1284488
#> 371  16.60814  21.40897 0.01554530 0.1284454
#> 372  16.59511  21.40087 0.01564311 0.1284422
#> 373  16.60829  21.40729 0.01553234 0.1284362
#> 374  16.59067  21.39648 0.01565851 0.1284254
#> 375  16.60818  21.40812 0.01550864 0.1284104
#> 376  16.59160  21.39618 0.01561871 0.1283941
#> 377  16.60196  21.40417 0.01552618 0.1283797
#> 378  16.59461  21.39874 0.01557187 0.1283696
#> 379  16.59673  21.40071 0.01554876 0.1283620
#> 380  16.59442  21.39891 0.01555862 0.1283545
#> 381  16.59325  21.39820 0.01555711 0.1283439
#> 382  16.59148  21.39632 0.01555639 0.1283305
#> 383  16.58906  21.39456 0.01556197 0.1283177
#> 384  16.59095  21.39462 0.01553586 0.1283079
#> 385  16.58611  21.39133 0.01556632 0.1283011
#> 386  16.59039  21.39355 0.01552782 0.1282965
#> 387  16.58337  21.38762 0.01557417 0.1282906
#> 388  16.58910  21.39060 0.01552131 0.1282831
#> 389  16.58182  21.38454 0.01556685 0.1282744
#> 390  16.58892  21.38871 0.01550342 0.1282664
#> 391  16.57888  21.38166 0.01557446 0.1282607
#> 392  16.59014  21.38705 0.01548220 0.1282581
#> 393  16.57149  21.37403 0.01562366 0.1282569
#> 394  16.59005  21.38560 0.01547685 0.1282541
#> 395  16.56693  21.36930 0.01564548 0.1282461
#> 396  16.58619  21.38215 0.01548382 0.1282326
#> 397  16.56862  21.37052 0.01560248 0.1282159
#> 398  16.58043  21.37842 0.01549539 0.1281997
#> 399  16.57110  21.37180 0.01555267 0.1281857
#> 400  16.57625  21.37610 0.01550147 0.1281730
#> 401  16.57203  21.37258 0.01552073 0.1281607
#> 402  16.57175  21.37190 0.01550989 0.1281485
#> 403  16.57137  21.37142 0.01550145 0.1281374
#> 404  16.56595  21.36666 0.01553359 0.1281295
#> 405  16.57333  21.37086 0.01547116 0.1281249
#> 406  16.56111  21.36200 0.01556151 0.1281223
#> 407  16.57269  21.36907 0.01546872 0.1281195
#> 408  16.56050  21.36022 0.01555552 0.1281135
#> 409  16.57163  21.36765 0.01546055 0.1281042
#> 410  16.55990  21.35828 0.01553640 0.1280921
#> 411  16.56901  21.36509 0.01545292 0.1280776
#> 412  16.55847  21.35687 0.01551791 0.1280632
#> 413  16.56842  21.36199 0.01542566 0.1280497
#> 414  16.55624  21.35306 0.01550551 0.1280370
#> 415  16.56375  21.35460 0.01543114 0.1280254
#> 416  16.55563  21.34845 0.01547789 0.1280107
#> 417  16.55600  21.34780 0.01546003 0.1279970
#> 418  16.55618  21.34619 0.01544602 0.1279869
#> 419  16.54952  21.34029 0.01548682 0.1279785
#> 420  16.55443  21.34132 0.01543833 0.1279711
#> 421  16.54759  21.33370 0.01548077 0.1279652
#> 422  16.55237  21.33414 0.01543543 0.1279607
#> 423  16.54285  21.32687 0.01550218 0.1279555
#> 424  16.55155  21.33279 0.01543058 0.1279507
#> 425  16.53672  21.32207 0.01553930 0.1279465
#> 426  16.55205  21.33301 0.01541970 0.1279438
#> 427  16.53034  21.31718 0.01558239 0.1279414
#> 428  16.55143  21.33163 0.01541781 0.1279385
#> 429  16.52690  21.31363 0.01559749 0.1279318
#> 430  16.54760  21.32751 0.01542354 0.1279170
#> 431  16.52779  21.31254 0.01555137 0.1278950
#> 432  16.54060  21.32229 0.01542953 0.1278700
#> 433  16.53126  21.31474 0.01547862 0.1278491
#> 434  16.53434  21.31605 0.01544198 0.1278373
#> 435  16.53523  21.31655 0.01543169 0.1278339
#> 436  16.52789  21.31054 0.01548659 0.1278340
#> 437  16.53800  21.31801 0.01540876 0.1278329
#> 438  16.52430  21.30939 0.01551022 0.1278281
#> 439  16.53903  21.31973 0.01538828 0.1278188
#> 440  16.52283  21.30923 0.01550180 0.1278072
#> 441  16.53519  21.31787 0.01539423 0.1277943
#> 442  16.52344  21.30956 0.01547208 0.1277823
#> 443  16.52864  21.31194 0.01541919 0.1277710
#> 444  16.52383  21.30851 0.01544526 0.1277604
#> 445  16.52382  21.30787 0.01543449 0.1277505
#> 446  16.52531  21.30792 0.01541364 0.1277425
#> 447  16.51946  21.30481 0.01545485 0.1277375
#> 448  16.52632  21.30676 0.01539430 0.1277335
#> 449  16.51519  21.29946 0.01547834 0.1277317
#> 450  16.52623  21.30605 0.01539107 0.1277305
#> 451  16.51100  21.29406 0.01550364 0.1277285
#> 452  16.52715  21.30406 0.01537435 0.1277246
#> 453  16.50895  21.29288 0.01550960 0.1277185
#> 454  16.52398  21.30212 0.01538402 0.1277096
#> 455  16.50893  21.29244 0.01548862 0.1276979
#> 456  16.52225  21.30196 0.01537439 0.1276853
#> 457  16.51033  21.29334 0.01545242 0.1276726
#> 458  16.51760  21.29810 0.01538493 0.1276612
#> 459  16.51385  21.29620 0.01540529 0.1276518
#> 460  16.51378  21.29577 0.01539814 0.1276446
#> 461  16.51579  21.29732 0.01537746 0.1276392
#> 462  16.50999  21.29289 0.01541701 0.1276348
#> 463  16.51674  21.29691 0.01536029 0.1276308
#> 464  16.50673  21.29052 0.01543421 0.1276273
#> 465  16.51700  21.29637 0.01535046 0.1276240
#> 466  16.50276  21.28790 0.01545877 0.1276213
#> 467  16.51628  21.29560 0.01534945 0.1276179
#> 468  16.50310  21.28640 0.01544581 0.1276134
#> 469  16.51325  21.29340 0.01536061 0.1276060
#> 470  16.50319  21.28699 0.01542911 0.1275966
#> 471  16.51015  21.29125 0.01536361 0.1275852
#> 472  16.50438  21.28772 0.01539790 0.1275747
#> 473  16.50822  21.29018 0.01536002 0.1275665
#> 474  16.50358  21.28712 0.01539050 0.1275612
#> 475  16.50478  21.28793 0.01537757 0.1275574
#> 476  16.50515  21.28838 0.01537215 0.1275546
#> 477  16.50229  21.28694 0.01539095 0.1275507
#> 478  16.50578  21.29004 0.01536011 0.1275456
#> 479  16.50021  21.28681 0.01539702 0.1275389
#> 480  16.50527  21.29099 0.01535243 0.1275322
#> 481  16.49924  21.28764 0.01539340 0.1275258
#> 482  16.50461  21.29137 0.01534735 0.1275208
#> 483  16.49669  21.28678 0.01540611 0.1275176
#> 484  16.50528  21.29223 0.01533888 0.1275169
#> 485  16.49228  21.28310 0.01544018 0.1275188
#> 486  16.50616  21.29280 0.01533889 0.1275238
#> 487  16.48832  21.27901 0.01547865 0.1275289
#> 488  16.50752  21.29373 0.01533731 0.1275325
#> 489  16.48536  21.27805 0.01550365 0.1275297
#> 490  16.50743  21.29551 0.01532760 0.1275195
#> 491  16.48504  21.27991 0.01548431 0.1275049
#> 492  16.50360  21.29335 0.01532824 0.1274901
#> 493  16.48942  21.28448 0.01542737 0.1274793
#> 494  16.49914  21.29024 0.01534432 0.1274721
#> 495  16.49288  21.28732 0.01538663 0.1274644
#> 496  16.49365  21.28744 0.01537003 0.1274543
#> 497  16.49705  21.29103 0.01533466 0.1274432
#> 498  16.49000  21.28659 0.01538248 0.1274363
#> 499  16.49851  21.29273 0.01531693 0.1274356
#> 500  16.48853  21.28667 0.01539761 0.1274386
#> 501  16.49861  21.29319 0.01531990 0.1274388
#> 502  16.48734  21.28625 0.01540172 0.1274331
#> 503  16.49592  21.29218 0.01532474 0.1274218
#> 504  16.48801  21.28710 0.01537312 0.1274090
#> 505  16.49416  21.29241 0.01531764 0.1273991
#> 506  16.48907  21.28983 0.01535216 0.1273933
#> 507  16.49275  21.29416 0.01532236 0.1273891
#> 508  16.49070  21.29386 0.01533444 0.1273839
#> 509  16.49112  21.29535 0.01532585 0.1273768
#> 510  16.49280  21.29679 0.01530505 0.1273685
#> 511  16.48841  21.29463 0.01533260 0.1273611
#> 512  16.49328  21.29782 0.01528945 0.1273555
#> 513  16.48634  21.29318 0.01533915 0.1273518
#> 514  16.49351  21.29865 0.01528213 0.1273490
#> 515  16.48398  21.29060 0.01535117 0.1273471
#> 516  16.49246  21.29766 0.01528624 0.1273454
#> 517  16.47903  21.28612 0.01538175 0.1273414
#> 518  16.49216  21.29568 0.01527946 0.1273389
#> 519  16.47648  21.28444 0.01539879 0.1273387
#> 520  16.49235  21.29421 0.01527550 0.1273387
#> 521  16.47440  21.28366 0.01541552 0.1273386
#> 522  16.49476  21.29741 0.01525635 0.1273358
#> 523  16.47399  21.28280 0.01540746 0.1273282
#> 524  16.49250  21.29784 0.01525546 0.1273147
#> 525  16.47759  21.28755 0.01535363 0.1272987
#> 526  16.48835  21.29403 0.01525436 0.1272832
#> 527  16.47863  21.28753 0.01531725 0.1272713
#> 528  16.48443  21.29007 0.01526324 0.1272639
#> 529  16.47910  21.28449 0.01529692 0.1272595
#> 530  16.47761  21.28380 0.01530563 0.1272562
#> 531  16.48239  21.28549 0.01526380 0.1272534
#> 532  16.47178  21.27797 0.01534267 0.1272512
#> 533  16.48346  21.28647 0.01525035 0.1272478
#> 534  16.46889  21.27535 0.01535793 0.1272452
#> 535  16.48254  21.28551 0.01525172 0.1272426
#> 536  16.46880  21.27711 0.01535657 0.1272405
#> 537  16.48128  21.28605 0.01525822 0.1272374
#> 538  16.46794  21.27678 0.01535612 0.1272331
#> 539  16.48164  21.28806 0.01524719 0.1272265
#> 540  16.46675  21.27664 0.01535181 0.1272188
#> 541  16.48072  21.28790 0.01523848 0.1272101
#> 542  16.46795  21.28020 0.01532956 0.1272017
#> 543  16.47819  21.28802 0.01524440 0.1271939
#> 544  16.47098  21.28514 0.01529727 0.1271886
#> 545  16.47524  21.28893 0.01526156 0.1271842
#> 546  16.47198  21.28759 0.01528394 0.1271804
#> 547  16.47313  21.28888 0.01527192 0.1271764
#> 548  16.47178  21.28780 0.01527707 0.1271715
#> 549  16.47125  21.28761 0.01527558 0.1271656
#> 550  16.47191  21.28790 0.01526408 0.1271594
#> 551  16.46992  21.28630 0.01527309 0.1271535
#> 552  16.47166  21.28881 0.01525624 0.1271481
#> 553  16.47013  21.28802 0.01526429 0.1271440
#> 554  16.47247  21.28981 0.01524343 0.1271409
#> 555  16.46819  21.28709 0.01527429 0.1271386
#> 556  16.47217  21.28940 0.01524121 0.1271367
#> 557  16.46590  21.28616 0.01528945 0.1271353
#> 558  16.47271  21.29111 0.01523714 0.1271348
#> 559  16.46308  21.28485 0.01531126 0.1271346
#> 560  16.47455  21.29473 0.01522687 0.1271352
#> 561  16.46019  21.28477 0.01533862 0.1271371
#> 562  16.47735  21.29837 0.01521276 0.1271401
#> 563  16.45513  21.28347 0.01538914 0.1271456
#> 564  16.47968  21.30054 0.01520581 0.1271502
#> 565  16.45029  21.27982 0.01543263 0.1271525
#> 566  16.47778  21.29873 0.01521478 0.1271453
#> 567  16.45030  21.27751 0.01540300 0.1271262
#> 568  16.47055  21.29305 0.01522194 0.1270981
#> 569  16.45647  21.28129 0.01529866 0.1270699
#> 570  16.46226  21.28538 0.01523549 0.1270509
#> 571  16.46283  21.28528 0.01522484 0.1270453
#> 572  16.45542  21.27783 0.01528253 0.1270497
#> 573  16.46651  21.28577 0.01520543 0.1270571
#> 574  16.45082  21.27365 0.01532811 0.1270614
#> 575  16.46790  21.28585 0.01519526 0.1270589
#> 576  16.44947  21.27244 0.01532622 0.1270496
#> 577  16.46559  21.28575 0.01519296 0.1270367
#> 578  16.45225  21.27600 0.01528193 0.1270242
#> 579  16.46040  21.28362 0.01521333 0.1270152
#> 580  16.45779  21.28352 0.01523019 0.1270096
#> 581  16.45602  21.28199 0.01523895 0.1270053
#> 582  16.46035  21.28748 0.01520455 0.1270005
#> 583  16.45395  21.28295 0.01524850 0.1269955
#> 584  16.46063  21.28795 0.01519455 0.1269922
#> 585  16.45252  21.28414 0.01525742 0.1269903
#> 586  16.46149  21.28985 0.01518693 0.1269893
#> 587  16.45139  21.28368 0.01526445 0.1269882
#> 588  16.46184  21.29203 0.01518246 0.1269847
#> 589  16.45088  21.28371 0.01525873 0.1269781
#> 590  16.45849  21.28985 0.01519301 0.1269694
#> 591  16.45404  21.28682 0.01521788 0.1269601
#> 592  16.45515  21.28793 0.01520219 0.1269524
#> 593  16.45534  21.28922 0.01519736 0.1269474
#> 594  16.45396  21.28884 0.01520579 0.1269444
#> 595  16.45567  21.29039 0.01519090 0.1269421
#> 596  16.45217  21.28958 0.01521778 0.1269398
#> 597  16.45597  21.29201 0.01518563 0.1269371
#> 598  16.44944  21.28801 0.01523224 0.1269329
#> 599  16.45535  21.29205 0.01518274 0.1269287
#> 600  16.44843  21.28817 0.01523194 0.1269236
#> 
get_metadata(amewoo, "is_sparse")
#> [1] FALSE