import BirdFlow model from an hdf5 file

This function imports a BirdFlow model data from an HDF5 file written by python.

Usage

import_birdflow(hdf5, ..., version)

Arguments

hdf5

Path to an hdf5 file

...

Arguments to be passed to a version specific internal function. Likely will only be used with version 1 which takes two additional arguments:

tiff

Path to the model geotiff

species

An eBird species code. It should appear in the species_code column of the data.frame returned by auk::get_ebird_taxonomy()

version

(optional) force reading of BirdFlow models as a particular version. Normally, this will be determined from metadata in the hdf5 file.

Value

a BirdFlow object

Details

The hdf5 file should have

date the date the model was exported from python

densities the training distributions in their vector forms stored in a matrix

dtuple (metadata list) with items:

cells the number of active (non-masked) cells

nan_mask a factor matrix of the raster extent with 'TRUE' for masked cells, and 'FALSE' for unmasked. This is the reverse of the BirdFlow internal representation which uses TRUE for active cells.

weeks the number of timesteps in the model

x_dim the number of columns in the model raster

y_dim the number of rows in the model raster

marginals the marginal probabilities between adjacent timesteps in the birdflow model. R reads these in column dominant format even though they are stored in row-dominant so they must be transposed.

version an integer indicating the version of the BirdFlow hdf5 file format. We will increment this when we change what's in the file.

The geoTiff is required for the coordinate reference system (CRS) and the spatial extent. In the future we hope to add this information, along with the species code to the hdf5 file. Currently the mask and some other attributes from the hdf5 are checked against redundant versions in the geoTiff.