Functions to convert among spatial indices

These functions allow conversions among different indices for locations in BirdFlow models and output. Specifically they convert among the raster spatial coordinates (x and y), raster row and col indices, and the index i along a location vector.

Usage

x_to_col(x, bf)

y_to_row(y, bf)

row_to_y(row, bf)

col_to_x(col, bf)

i_to_rc(i, bf)

i_to_row(i, bf)

i_to_col(i, bf)

i_to_x(i, bf)

i_to_y(i, bf)

i_to_xy(i, bf)

rc_to_i(row, col, bf)

xy_to_i(x, y, bf)

Arguments

x, y

The x and y coordinates associated with the BirdFlow model or data. This will typically be an easting and northing in meters. A two column matrix containing x and y columns can also be passed to x in which case y should be omitted.

bf

a BirdFlow model

row, col

The row and column index of a cell in the BirdFlow model and associated raster data. Alternatively, a two column matrix containing row and column indices in columns 1 and 2 respectively can be passed to row in which case col should be omitted.

i

The index along a state vector that contains the data for unmasked cells.

Value

x_to_col(x, bf) and y_to_row(y, bf)

return the column or row index that each x or y coordinate falls within.

row_to_y(row, bf) and col_to_x(col, bf)

return the y or x coordinate of the center each row or column.

i_to_row(i, bf) and i_to_col(i, bf)

return the row or column index corresponding to the vector state index, i.

i_to_rc(i, bf)

returns a two column matrix of the row and column index in the raster corresponding to the index, i of the vector state.

i_to_x(i, bf) and i_to_y(i, bf)

return the x or y coordinate from the vector state index, i.

i_to_xy(i, bf)

returns a two column matrix of the x and y coordinates corresponding to the index, i of the vector state space.

rc_to_i(row, col, bf) and xy_to_i(x, y, bf)

return the state space index corresponding to x and y coordinates or row and column indices.

See also