Lookup a sequence of timesteps

lookup_timestep_sequence() returns an ordered vector of timesteps possibly crossing over the year boundary.

Usage

lookup_timestep_sequence(x, start, end, direction, season_buffer)

Arguments

x

A BirdFlow object

start, end

The starting and ending points in time specified as timesteps, character dates, date objects, or if end is omitted start may be set to "all" or a season name to be interpreted by lookup_season_timesteps().

direction

Either "forward" or "backward".

If start and end represent dates and direction is used an error will be thrown if direction isn't consistent with direction indicated by the dates.

If start and end are not dates, direction defaults to "forward" and start and end should either both be timesteps (numeric); or end should be omitted and start should be "all" or a season name.

season_buffer

Only used if start is a season. season_buffer is passed to lookup_season_timesteps() and defaults to 1; it is the number of timesteps to extend the season by at each end.

Value

An integer sequence of timesteps.

Details

start and end will fall into one of three categories which are treated differently:

If start and end are numeric than they are treated as timesteps and direction is used to route either forward (default) or backwards between them, possibly passing over the year boundary.

If start and end are date objects or both characters they are treated as dates, and the direction is determined from the dates:"forward" if in chronological order, "backwards" otherwise. If the direction argument is used along with date input and it is not consistent with the implicit direction from the dates, than an error will be thrown.

If start is a character and end is missing than it is assumed that start is a key word that is either "all" for all timesteps, or a season name. Season names are passed to lookup_season_timesteps() along with season_buffer. direction will be followed and will default to "forward".

Examples

bf <- BirdFlowModels::rewbla

# Using timesteps - direction defaults to "forward"
lookup_timestep_sequence(bf, 50, 3)  # c(50:52, 1:3)
#> [1] 50 51 52  1  2  3
lookup_timestep_sequence(bf, 50, 3, direction = "backward") # 50:3
#>  [1] 50 49 48 47 46 45 44 43 42 41 40 39 38 37 36 35 34 33 32 31 30 29 28 27 26
#> [26] 25 24 23 22 21 20 19 18 17 16 15 14 13 12 11 10  9  8  7  6  5  4  3

# Using dates - order of dates determines direction
lookup_timestep_sequence(bf, "2023-12-1", "2024-01-20")
#> [1] 48 49 50 51 52  1  2  3
lookup_timestep_sequence(bf, "2024-01-20", "2023-12-1")
#> [1]  3  2  1 52 51 50 49 48

# "all" - direction defaults to "forward"
lookup_timestep_sequence(bf, "all")
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50
#> [51] 51 52
lookup_timestep_sequence(bf, "all", direction = "backward")
#>  [1] 52 51 50 49 48 47 46 45 44 43 42 41 40 39 38 37 36 35 34 33 32 31 30 29 28
#> [26] 27 26 25 24 23 22 21 20 19 18 17 16 15 14 13 12 11 10  9  8  7  6  5  4  3
#> [51]  2  1

# Season - direction defaults to "forward", season_buffer defaults to 1
lookup_timestep_sequence(bf, "prebreeding_migration")
#>  [1]  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18
lookup_timestep_sequence(bf, "prebreeding_migration", season_buffer = 0,
                         direction = "backward")
#>  [1] 17 16 15 14 13 12 11 10  9  8  7  6  5