prepare eBird Status and Trends data for BirdFlow model fitting

Write a template BirdFlow object to an hdf5 file based on distribution data downloaded with ebirdst. The object is complete except for marginals and transitions.

Usage

preprocess_species(
  species,
  out_dir,
  res = NULL,
  hdf5 = TRUE,
  tiff = FALSE,
  overwrite = TRUE,
  crs,
  clip,
  max_params,
  gpu_ram = 12,
  skip_quality_checks = FALSE,
  dummy_dynamic_mask = FALSE
)

Arguments

species

a species in any format accepted by ebirdst::get_species()

out_dir

output directory, files will be written here. Required unless both tiff and hdf5 are TRUE. File names created here will incorporate the species code, resolution, and eBird version year.

res

the target resolution of the BirdFlow model in kilometers. If res is NULL (default) then a resolution that results in less than max_params parameters will be used, while also minimizing the resolution and limiting the number of significant digits.

hdf5

if TRUE (default) an hdf5 file will be exported.

tiff

set to TRUE to export geoTIFF files. They are not needed to fit the BirdFlow model.

overwrite

if TRUE (default) any pre-existing output files will be overwritten. If FALSE pre-existing files will result in an error.

crs

coordinate reference system (CRS) to use. Defaults to the custom projection eBird has assigned to this species - see ebirdst::load_fac_map_parameters()). It will be interpreted by terra::crs() to generate a well known text representation of the CRS.

clip

a polygon or the path to a file containing a polygon. It must have a CRS and should either be a SpatVector() object or or produce one when called with vect(clip)

max_params

the maximum number of fitted parameters that the BirdFlow model should contain. Ignored if res is not NULL. Otherwise a resolution will be chosen that yields this many fitted parameters. See gpu_ram for the default way of setting max_params and res.

gpu_ram

Gigabytes of ram on GPU machine that will fit the models. If res is NULL and max_params is missing this is used to estimate max_paramswhich is, in turn, used to determine the resolution. Ignored if either res or max_params is set.

skip_quality_checks

If TRUE than preprocess the species even if not all of four ranges are modeled (based on ebirdst_runs()).

dummy_dynamic_mask

set to TRUE to force all cells in the dynamic mask to TRUE. This creates a BirdFlow object with a dynamic mask component that is compatible with the current BirdFlowR package but that mimics the older predynamic mask BirdFlow models. If fit the resulting object will have transitions at every timestep among all active cells. The only reason to set this to TRUE is for comparsion testing and quality control during the transition, and this parameter may eventually be dropped.

Value

returns a BirdFlow model object that lacks marginals, but is otherwise complete.

Maximum number of parameters

The maximum number of parameters that can be fit is machine dependent. 2023-02-10 we tested under different resolutions with "amewoo" and identified bounds on the maximum.

Machine	GPU Ram (GB)	Lower Bound (worked)	Upper Bound (failed)	Params / GB

titanx gpu	12GB	306804561	334693725	25567047

m40 gpu	24GB	557395226	610352178	23224801

The number of parameters is the number of unmasked cells for the first timestep + the total number of cells in the marginals which is calculated from the dynamic mask.

If gpu_ram is used (and not res or max_parameters ) than max_parameters is set to 23,224,801 * gpu_ram (lower of two values in table above).

The heuristic to determine resolution given a maximum number of parameters must estimate the number of cells covered by the data at a different resolution, a noisy process, so it iteratively tries to find the smallest resolution that doesn't exceed max_params and then rounds to a slightly larger resolution (fewer parameters).

Examples

if (FALSE) {

 bf <- preprocess_species("amewoo", tiff = FALSE, hdf5 = FALSE )
 plot(rasterize_distr(get_distr( bf, c(1, 26)), bf))

# Create clip polgyon as an sf object
# Use the extent rectangle but with western edge moved in
# The clip can be anything that terra::vect will process into a polygon
e <- ext(bf)
e[1] <- -1500000
coords <- matrix(c(e[1], e[3],
                   e[1], e[4],
                   e[2], e[4],
                   e[2], e[3],
                   e[1], e[3]), ncol = 2, byrow = TRUE)
sfc <- st_sfc(st_polygon(list(coords)), crs = crs(bf))
clip <- st_sf(data.frame(id = 1, geom = sfc))

bfc <- preprocess_species("amewoo", tiff = FALSE,
                         hdf5 = FALSE, clip = clip ) # clipped bird flow

 plot(rasterize_distr(get_distr(bfc, 1), bfc))

}